Preserve datapath when GlobalnetPod is restarted
In the current implementation of Globalnet, when the Globalnet Pod
is restarted, the datapath goes down as Globalnet controller flushes
the IPTable chains that support inter-cluster connectivity.

In this PR we move the cleanup to RouteAgent Pod which does the cleanup
(aka flushing of iptable chains) when there is a Gateway migration.

This PR also moves the IPTable Chain names to routeAgent code to avoid
cyclic dependency and to have all the chain names at a single place.

Fixes Issue: https://github.com/submariner-io/submariner/issues/721

Signed-Off-by: Sridhar Gaddam <sgaddam@redhat.com>